refactor(slash-command): move interception from agent to web layer
Slash commands are a routing concern, not a domain concern. Move the
interception logic out of Agents::send_message into the web layer's
api_agent_message handler so the agent module stays clean.

- Remove slash_commands field and try_slash_command from Agents
- Add SlashCommands::is_slash_command() and process() methods
- App holds SlashCommands and exposes via slash_commands() accessor
- Web handler routes: if slash → app.slash_commands().process()
  else → app.agents().send_message()
